[EarlyIfConverter] Move SSAIfConv into a spearate header/source
